Other CQL functions (unimplemented)

These CQL commands are not yet implemented in stratigraphr and only included here for reference. Unimplemented commands can be included manually using cql() (see examples), but if you need one for an analysis please file a bug report!

cql_age(...)

cql_axis(...)

cql_c_combine(...)

cql_c_simulate(...)

cql_correl_matrix(...)

cql_correlation(...)

cql_covar_matrix(...)

cql_curve(...)

cql_delta_r(...)

cql_difference(...)

cql_end(...)

cql_exp(...)

cql_gap(...)

cql_interval(...)

cql_kde_model(...)

cql_kde_plot(...)

cql_label(...)

cql_line(...)

cql_mcmc_sample(...)

cql_mix_curves(...)

cql_number(...)

cql_offset(...)

cql_outlier(...)

cql_outlier_model(...)

cql_p(...)

cql_pois(...)

cql_prior(...)

cql_probability(...)

cql_r_combine(...)

cql_r_simulate(...)

cql_reservoir(...)

cql_sample(...)

cql_sapwood(...)

cql_sapwood_model(...)

cql_shift(...)

cql_start(...)

cql_after()

cql_before()

cql_combine()

cql_first()

cql_last()

cql_order()

cql_span()

cql_sum()

Arguments

...
Ignored.

Value

A warning not to use these functions!

See also

Examples

# Use cql() to manually write commands that are not yet implemented, e.g.:
cql("After(\"A\");")

#>  // CQL2 generated by stratigraphr v0.3.0
#>  After("A");