stratigraphr: Archaeological Stratigraphy and Chronological Sequences

A tidy framework for working with archaeological stratigraphy and chronology. Includes tools for reading, analysing, and visualising stratigraphy (Harris matrices) and sequences as directed graphs (Dye & Buck 2015) doi:10.1016/j.jas.2015.08.008 ; helper functions for using radiocarbon dates in a tidy data analysis; and an R interface to the Chronological Query Language (CQL) used by 'OxCal' (Bronk Ramsey 2009) doi:10.1017/S0033822200033865 .

See also

Author

Maintainer: Joe Roe joe@joeroe.io (ORCID)