Calibrate radiocarbon dates

A thin wrapper of rcarbon::calibrate() that returns calibrated dates as a single list rather than a `CalDates`` object. Useful, for example, if you want to add a column of calibrated dates to an existing table with dplyr::mutate()

c14_calibrate(cra, error, ..., engine = c("rcarbon", "OxCal", "Bchron"))

Arguments

cra

A vector of uncalibrated radiocarbon ages.

error

A vector of standard errors associated with cra

...

Optional arguments passed to calibration function.

engine

Package to use for calibration, i.e. rcarbon::calibrate() ("rcarbon"), oxcAAR::oxcalCalibrate() ("OxCal"), or Bchron::BchronCalibrate() ("Bchron"). Default: "rcarbon".

Value

A list of cal objects.

See also

Examples

data("shub1_radiocarbon")
shub1_radiocarbon %>%
  dplyr::mutate(cal = c14_calibrate(cra, error))
#> Warning: stratigraphr::c14_calibrate() is deprecated. Please use c14::c14_calibrate() instead.
#> [1] "Calibrating radiocarbon ages..."
#>   |                                                                              |                                                                      |   0%  |                                                                              |===                                                                   |   4%  |                                                                              |=====                                                                 |   7%  |                                                                              |========                                                              |  11%  |                                                                              |==========                                                            |  15%  |                                                                              |=============                                                         |  19%  |                                                                              |================                                                      |  22%  |                                                                              |==================                                                    |  26%  |                                                                              |=====================                                                 |  30%  |                                                                              |=======================                                               |  33%  |                                                                              |==========================                                            |  37%  |                                                                              |=============================                                         |  41%  |                                                                              |===============================                                       |  44%  |                                                                              |==================================                                    |  48%  |                                                                              |====================================                                  |  52%  |                                                                              |=======================================                               |  56%  |                                                                              |=========================================                             |  59%  |                                                                              |============================================                          |  63%  |                                                                              |===============================================                       |  67%  |                                                                              |=================================================                     |  70%  |                                                                              |====================================================                  |  74%  |                                                                              |======================================================                |  78%  |                                                                              |=========================================================             |  81%  |                                                                              |============================================================          |  85%  |                                                                              |==============================================================        |  89%  |                                                                              |=================================================================     |  93%  |                                                                              |===================================================================   |  96%  |                                                                              |======================================================================| 100%
#> [1] "Done."
#> Warning: stratigraphr::as_cal() is deprecated. Please use c14::as_cal() instead.
#> # A tibble: 27 × 9
#>    lab_id      context phase   sample          mater…¹   cra error outlier cal  
#>    <chr>         <int> <chr>   <chr>           <chr>   <int> <int> <lgl>   <lis>
#>  1 RTD-7951         23 Phase 7 Context 126, S… Bolbos… 12166    55 FALSE   <cal>
#>  2 Beta-112146      24 Phase 7 SHUB1/105       gazell… 12310    60 FALSE   <cal>
#>  3 RTD-7317         26 Phase 7 Context 83, Sa… Bolbos… 12289    46 FALSE   <cal>
#>  4 RTD-7318         27 Phase 7 Context 86, Sa… Zilla … 12332    46 FALSE   <cal>
#>  5 RTD-7948         24 Phase 7 Context 120, S… Bolbos… 12478    38 FALSE   <cal>
#>  6 RTD-7947         22 Phase 6 Context 166, K… Bolbos… 12322    38 FALSE   <cal>
#>  7 RTD-7313         22 Phase 6 Context 67, K2… Bolbos… 12346    46 FALSE   <cal>
#>  8 RTD-7311         22 Phase 6 Context 67, K2… Vitex … 12367    65 FALSE   <cal>
#>  9 RTD-7312         22 Phase 6 Context 67, K2… Zilla … 12405    50 FALSE   <cal>
#> 10 RTD-7314         22 Phase 6 Context 67, K2… Bolbos… 12273    48 FALSE   <cal>
#> # … with 17 more rows, and abbreviated variable name ¹​material