Calibrate radiocarbon dates with tidy syntax

A thin wrapper of rcarbon::calibrate() that returns calibrated dates as a single list rather than a `CalDates`` object. Useful, for example, if you want to add a column of calibrated dates to an existing table with dplyr::mutate()

cal(cra, error, ...)

Arguments

cra
A vector of uncalibrated radiocarbon ages.

error
A vector of standard errors associated with cra

...
Optional arguments passed to rcarbon::calibrate()

Value

A list of calGrid objects. See rcarbon::calibrate() for details.

Note

This function only supports the 'grid' method of representing calibrated dates (i.e. calMatrix is always set to FALSE) and the metadata usually returned by rcarbon::calibrate() is discarded.

Examples

data("shub1_radiocarbon")
shub1_radiocarbon %>%
  dplyr::mutate(CalDate = cal(cra, error, normalise = FALSE, verbose = FALSE))

#> Registered S3 method overwritten by 'cli':
#>   method     from    
#>   print.boxx spatstat

#> # A tibble: 27 x 9
#>    lab_id   context phase  sample       material      cra error outlier CalDate 
#>    <chr>      <int> <chr>  <chr>        <chr>       <int> <int> <lgl>   <named >
#>  1 RTD-7951      23 Phase… Context 126… Bolboschoe… 12166    55 FALSE   <data.f…
#>  2 Beta-11…      24 Phase… SHUB1/105    gazelle du… 12310    60 FALSE   <data.f…
#>  3 RTD-7317      26 Phase… Context 83,… Bolboschoe… 12289    46 FALSE   <data.f…
#>  4 RTD-7318      27 Phase… Context 86,… Zilla sp. … 12332    46 FALSE   <data.f…
#>  5 RTD-7948      24 Phase… Context 120… Bolboschoe… 12478    38 FALSE   <data.f…
#>  6 RTD-7947      22 Phase… Context 166… Bolboschoe… 12322    38 FALSE   <data.f…
#>  7 RTD-7313      22 Phase… Context 67,… Bolboschoe… 12346    46 FALSE   <data.f…
#>  8 RTD-7311      22 Phase… Context 67,… Vitex sp. … 12367    65 FALSE   <data.f…
#>  9 RTD-7312      22 Phase… Context 67,… Zilla sp. … 12405    50 FALSE   <data.f…
#> 10 RTD-7314      22 Phase… Context 67,… Bolboschoe… 12273    48 FALSE   <data.f…
#> # … with 17 more rows