Describe a boundary constraint in CQL

The CQL command Boundary describes constraints within an ordered sequence (see cql_sequence()). Groups of events between two boundaries are assumed to be sampled from the same prior distribution. Boundary alone models a uniform prior. Other types of boundary, i.e. Sigma_Boundary, Tau_Boundary, and Zero_Boundary are not yet implemented in stratigraphr.

Boundaries can contain a Transition command, in which case they describe a non-instantaneous transition or 'trapezium' prior.

cql_boundary(name, ..., prior = NULL)

cql_sigma_boundary()

cql_tau_boundary()

cql_zero_boundary()

cql_transition(name, prior = NULL)

Arguments

name
Character. Label for a boundary or transition.

...
cql objects contained within a boundary.

prior
cql object. Expression describing the prior likelihood of a boundary or transition.

Value

A cql object.

References

https://c14.arch.ox.ac.uk/oxcalhelp/hlp_analysis_oper.html#group

https://c14.arch.ox.ac.uk/oxcalhelp/hlp_commands.html

See also

Examples

# Uniform model with unknown boundaries
cql(
  cql_boundary("P1 start"),
  cql_phase("P1",
    cql_r_date("A", 5050, 30),
    cql_r_date("B", 5000, 30),
    cql_r_date("C", 4950, 30)
  ),
  cql_boundary("P1 end")
)

#>  // CQL2 generated by stratigraphr v0.2.0
#>  Boundary("P1 start");
#>  Phase("P1")
#>  {
#>   R_Date("A", 5050, 30);
#>   R_Date("B", 5000, 30);
#>   R_Date("C", 4950, 30);
#>  };
#>  Boundary("P1 end");

# Uniform model with boundaries with a prior likelihood
cql(
  cql_boundary("P1 start",
               prior = cql_date("P1S-Prior", cql_u("", 5200, 5100))),
  cql_phase("P1",
    cql_r_date("A", 5050, 30),
    cql_r_date("B", 5000, 30),
    cql_r_date("C", 4950, 30)
  ),
  cql_boundary("P1 end",
               prior = cql_date("P1E-Prior", cql_u("", 4800, 4900)))
)

#>  // CQL2 generated by stratigraphr v0.2.0
#>  Boundary("P1 start", Date("P1S-Prior", U("", 5200, 5100)));
#>  Phase("P1")
#>  {
#>   R_Date("A", 5050, 30);
#>   R_Date("B", 5000, 30);
#>   R_Date("C", 4950, 30);
#>  };
#>  Boundary("P1 end", Date("P1E-Prior", U("", 4800, 4900)));

# 'Trapezium' model
cql(
  cql_boundary("P1 start",
    cql_transition("")
  ),
  cql_phase("P1",
    cql_r_date("A", 5050, 30),
    cql_r_date("B", 5000, 30),
    cql_r_date("C", 4950, 30)
  ),
  cql_boundary("P1 end",
    cql_transition("")
  )
)

#>  // CQL2 generated by stratigraphr v0.2.0
#>  Boundary("P1 start")
#>  {
#>   Transition("");
#>  };
#>  Phase("P1")
#>  {
#>   R_Date("A", 5050, 30);
#>   R_Date("B", 5000, 30);
#>   R_Date("C", 4950, 30);
#>  };
#>  Boundary("P1 end")
#>  {
#>   Transition("");
#>  };